Added the 5 buttons that are to be on the bottom to the window. Added IDS to wxIDS.h for the new Buttons.
New Issue
Summary: Bottom Buttons do not stay alligned if the strings in the tabs change.
Labels: Priority-Low
The bottom 5 buttons (Close, Help, etc) do not stay alligned with the
tabs for changing pages in the launcher, because the buttons' spacing from
the left side of the window is hard coded.