Just let CreateTempFileName find the tempdir by itself
If we don't give CreateTempFileName a patch it doesn't try to split
it and use the last component of the path as a prefix.  This means
that temp directories that are only one off of root (like C:\TMP)
will now work.
Also lets CreateTempFileName open the file for us as well. Its less
racy that way.