Fix first run on Linux (#146)
wxLauncher failed to start the game because MigrateOldConfig failed.
When no new config was present, it assumed that the old config must be
copied over to the new location. On the first run, there was no old
config, so MigrateOldConfig failed and returned false, meaning the game
was not started.

First check for the existance of the old config file and only try to
copy it if exists.